# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IP Info by Shifter",
    "version": "1.0.0",
    "description": "Free public IP geolocation and ASN API. Maintained and supported by Shifter. No signup or API key.",
    "termsOfService": "https://ip-info.com/terms",
    "contact": {
      "name": "Shifter",
      "url": "https://shifter.io",
      "email": "hi@shifter.io"
    }
  },
  "servers": [
    {
      "url": "https://ip-info.com"
    },
    {
      "url": "http://ip-info.com"
    }
  ],
  "security": [],
  "paths": {
    "/myip": {
      "get": {
        "operationId": "getMyIp",
        "summary": "Return only the caller IP as plain text",
        "description": "Returns the normalized IPv4 or IPv6 address followed by a newline. Uses the same trusted-proxy and authenticated-ingress rules as /json. Query parameters are ignored. No database is required; local/private caller addresses are returned too. Errors use the standard JSON envelope.",
        "security": [],
        "responses": {
          "200": {
            "description": "Caller IP address followed by a newline; no JSON wrapper.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "8.8.8.8\n"
              }
            }
          },
          "400": {
            "description": "Unable to determine the caller IP, or invalid/missing trusted-ingress headers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_ip",
                    "message": "Malformed, conflicting, non-public, or unsupported lookup input."
                  }
                }
              }
            }
          },
          "405": {
            "description": "Use GET for lookups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "method_not_allowed",
                    "message": "Use GET for lookups."
                  }
                }
              }
            }
          },
          "431": {
            "description": "Request headers exceed the configured limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "request_too_large",
                    "message": "Request headers exceed the configured limit."
                  }
                }
              }
            }
          },
          "504": {
            "description": "Request timed out; retry later with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "timeout",
                    "message": "Request timed out; retry later with backoff."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/json": {
      "get": {
        "operationId": "lookupIp",
        "summary": "Look up the caller or a supplied IP",
        "description": "Free, no API key. Caller detection returns the requesting machine or proxy exit, not necessarily the human user's IP.",
        "security": [],
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Public IPv4 or IPv6 literal. Omit to detect caller. Must match path IP when both are supplied.",
            "example": "8.8.8.8"
          }
        ],
        "responses": {
          "200": {
            "description": "IP lookup result; missing attributes are null.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpInfo"
                },
                "example": {
                  "ip": "8.8.8.8",
                  "city": "Mountain View",
                  "city_geoname_id": 5375480,
                  "region": "California",
                  "region_code": "CA",
                  "region_geoname_id": 5332921,
                  "district": "Santa Clara",
                  "district_code": null,
                  "district_geoname_id": 5393021,
                  "country": "US",
                  "country_name": "United States",
                  "country_geoname_id": 6252001,
                  "is_eu": false,
                  "continent": "NA",
                  "continent_name": "North America",
                  "continent_geoname_id": 6255149,
                  "loc": "37.422,-122.085",
                  "latitude": 37.422,
                  "longitude": -122.085,
                  "postal": "94043",
                  "timezone": "America/Los_Angeles",
                  "weather_code": "USCA0746",
                  "asn": 15169,
                  "as_name": "Google LLC",
                  "isp": "Google LLC",
                  "org": "Level 3",
                  "connection_type": "Corporate",
                  "user_type": "hosting",
                  "is_anycast": true
                }
              }
            }
          },
          "400": {
            "description": "Malformed, conflicting, non-public, or unsupported lookup input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_ip",
                    "message": "Malformed, conflicting, non-public, or unsupported lookup input."
                  }
                }
              }
            }
          },
          "404": {
            "description": "No record in the active database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "ip_not_found",
                    "message": "No record in the active database."
                  }
                }
              }
            }
          },
          "405": {
            "description": "Use GET for lookups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "method_not_allowed",
                    "message": "Use GET for lookups."
                  }
                }
              }
            }
          },
          "431": {
            "description": "Request headers exceed the configured limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "request_too_large",
                    "message": "Request headers exceed the configured limit."
                  }
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry later with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "database_unavailable",
                    "message": "Database unavailable; retry later with backoff."
                  }
                }
              }
            }
          },
          "504": {
            "description": "Request timed out; retry later with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "timeout",
                    "message": "Request timed out; retry later with backoff."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{ip}/json": {
      "get": {
        "operationId": "lookupIpByPath",
        "summary": "Look up a supplied IP",
        "description": "Free, no API key. Caller detection returns the requesting machine or proxy exit, not necessarily the human user's IP.",
        "security": [],
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Public IPv4 or IPv6 literal.",
            "example": "8.8.8.8"
          },
          {
            "name": "ip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Public IPv4 or IPv6 literal. Omit to detect caller. Must match path IP when both are supplied.",
            "example": "8.8.8.8"
          }
        ],
        "responses": {
          "200": {
            "description": "IP lookup result; missing attributes are null.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpInfo"
                },
                "example": {
                  "ip": "8.8.8.8",
                  "city": "Mountain View",
                  "city_geoname_id": 5375480,
                  "region": "California",
                  "region_code": "CA",
                  "region_geoname_id": 5332921,
                  "district": "Santa Clara",
                  "district_code": null,
                  "district_geoname_id": 5393021,
                  "country": "US",
                  "country_name": "United States",
                  "country_geoname_id": 6252001,
                  "is_eu": false,
                  "continent": "NA",
                  "continent_name": "North America",
                  "continent_geoname_id": 6255149,
                  "loc": "37.422,-122.085",
                  "latitude": 37.422,
                  "longitude": -122.085,
                  "postal": "94043",
                  "timezone": "America/Los_Angeles",
                  "weather_code": "USCA0746",
                  "asn": 15169,
                  "as_name": "Google LLC",
                  "isp": "Google LLC",
                  "org": "Level 3",
                  "connection_type": "Corporate",
                  "user_type": "hosting",
                  "is_anycast": true
                }
              }
            }
          },
          "400": {
            "description": "Malformed, conflicting, non-public, or unsupported lookup input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "invalid_ip",
                    "message": "Malformed, conflicting, non-public, or unsupported lookup input."
                  }
                }
              }
            }
          },
          "404": {
            "description": "No record in the active database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "ip_not_found",
                    "message": "No record in the active database."
                  }
                }
              }
            }
          },
          "405": {
            "description": "Use GET for lookups.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "method_not_allowed",
                    "message": "Use GET for lookups."
                  }
                }
              }
            }
          },
          "431": {
            "description": "Request headers exceed the configured limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "request_too_large",
                    "message": "Request headers exceed the configured limit."
                  }
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable; retry later with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "database_unavailable",
                    "message": "Database unavailable; retry later with backoff."
                  }
                }
              }
            }
          },
          "504": {
            "description": "Request timed out; retry later with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "timeout",
                    "message": "Request timed out; retry later with backoff."
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}